pyproject.toml
[project]
name = "aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.15"
dependencies = [
    "boto3-stubs[all]>=1.39.4",
    "fastapi[all]>=0.116.1",
    "groq>=0.30.0",
    "moto>=5.1.8",
    "pydantic[all]>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pylint>=3.3.7",
    "pytest>=8.4.1",
]
